feat: implement folder structure and controller modularization
