+++ src/commonMain/kotlin/main.kt
@@ -1,70 +1,24 @@
-import com.soywiz.klock.seconds
 import com.soywiz.korge.*
-import com.soywiz.korge.tween.*
-import com.soywiz.korge.view.*
+import com.soywiz.korge.scene.Module
 import com.soywiz.korim.color.*
-import com.soywiz.korim.font.readBitmapFont
-import com.soywiz.korim.format.*
-import com.soywiz.korim.text.TextAlignment
-import com.soywiz.korio.file.std.*
-import com.soywiz.korma.geom.Rectangle
-import com.soywiz.korma.geom.degrees
-import com.soywiz.korma.geom.vector.roundRect
-import com.soywiz.korma.interpolation.Easing
-import storage.Storage
+import com.soywiz.korinject.AsyncInjector
+import com.soywiz.korma.geom.SizeInt
+import controller.GameScene
+import controller.MenuController
 
-suspend fun main() = Korge(width = 512, height = 824, bgcolor = RGBA(253, 247, 240)) {
-    //val storage = Storage()
-    //val user = storage.getUser()
-    //println(user)
-    val font = resourcesVfs["clear_sans.fnt"].readBitmapFont()
-    val cellSize = views.virtualWidth / 5.0
-    val fieldSize = 50 + 4 * cellSize
-    val leftIndent = (views.virtualWidth - fieldSize) / 2
-    val topIndent = 150.0
-    val bgField = roundRect(fieldSize, fieldSize, 5.0, fill = Colors["#b9aea0"]) {
-        position(leftIndent, topIndent)
-    }
-    graphics {
-        position(leftIndent, topIndent)
-        fill(Colors["#cec0b2"]) {
-            for (i in 0..3) {
-                for (j in 0..3) {
-                    roundRect(10 + (10 + cellSize) * i, 10 + (10 + cellSize) * j, cellSize, cellSize, 5.0)
-                }
-            }
-        }
-    }
-    val bgLogo = roundRect(cellSize, cellSize, 5.0, fill = Colors["#edc403"]) {
-        position(leftIndent, 30.0)
-    }
-    text("2048", cellSize * 0.5, Colors.WHITE, font).centerOn(bgLogo)
-    val bgBest = roundRect(cellSize * 1.5, cellSize * 0.8, 5.0, fill = Colors["#bbae9e"]) {
-        alignRightToRightOf(bgField)
-        alignTopToTopOf(bgLogo)
-    }
-    val bgScore = roundRect(cellSize * 1.5, cellSize * 0.8, 5.0, fill = Colors["#bbae9e"]) {
-        alignRightToLeftOf(bgBest, 24)
-        alignTopToTopOf(bgBest)
-    }
-    text("BEST", cellSize * 0.25, RGBA(239, 226, 210), font) {
-        centerXOn(bgBest)
-        alignTopToTopOf(bgBest, 5.0)
-    }
-    text("0", cellSize * 0.5, Colors.WHITE, font) {
-        setTextBounds(Rectangle(0.0, 0.0, bgBest.width, cellSize - 24.0))
-        alignment = TextAlignment.MIDDLE_CENTER
-        alignTopToTopOf(bgBest, 12.0)
-        centerXOn(bgBest)
-    }
-    text("SCORE", cellSize * 0.25, RGBA(239, 226, 210), font) {
-        centerXOn(bgScore)
-        alignTopToTopOf(bgScore, 5.0)
-    }
-    text("0", cellSize * 0.5, Colors.WHITE, font) {
-        setTextBounds(Rectangle(0.0, 0.0, bgScore.width, cellSize - 24.0))
-        alignment = TextAlignment.MIDDLE_CENTER
-        centerXOn(bgScore)
-        alignTopToTopOf(bgScore, 12.0)
+suspend fun main() = Korge(Korge.Config(module = GameModule))
+
+object GameModule : Module() {
+    override val mainScene = MenuController::class
+    override val bgcolor = RGBA(253, 247, 240)
+    override val size = SizeInt(1080, 1920)
+    override val windowSize = SizeInt(1080, 1920)
+
+    override suspend fun AsyncInjector.configure() {
+        mapInstance(Dependency("HELLO WORLD"))
+        mapPrototype { GameScene(get()) }
+        mapPrototype { MenuController(get()) }
     }
 }
+
+class Dependency(val value: String)
