+++ src/commonMain/kotlin/model/User.kt
@@ -1,0 +1,4 @@
+package model
+
+data class User(val id: String) {
+}
